requirements.txt
langchain
langchain-community
langchain-openai
langchain_voyageai
langchain_pinecone
langchain-anthropic
langchain-chroma
langchain-pinecone
langchain-voyageai
streamlit
st-pages
streamlit-authenticator
pinecone
tiktoken
pypdf
python-dotenv
pandas
matplotlib
PyPDF2
lark
openpyxl
bs4
unstructured
fastapi
uvicorn